Nest dialog class and apply NRT
