+++ osu.Game/Online/Chat/ExternalLinkOpener.cs
@@ -1,27 +1,27 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
+using System;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
+using osu.Framework.Graphics.Sprites;
 using osu.Framework.Platform;
 using osu.Game.Configuration;
 using osu.Game.Overlays;
-using osu.Game.Overlays.Chat;
+using osu.Game.Overlays.Dialog;
 
 namespace osu.Game.Online.Chat
 {
     public class ExternalLinkOpener : Component
     {
         [Resolved]
-        private GameHost host { get; set; }
+        private GameHost host { get; set; } = null!;
 
         [Resolved(CanBeNull = true)]
-        private IDialogOverlay dialogOverlay { get; set; }
+        private IDialogOverlay? dialogOverlay { get; set; }
 
-        private Bindable<bool> externalLinkWarning;
+        private Bindable<bool> externalLinkWarning = null!;
 
         [BackgroundDependencyLoader(true)]
         private void load(OsuConfigManager config)
@@ -31,10 +31,39 @@ namespace osu.Game.Online.Chat
 
         public void OpenUrlExternally(string url, bool bypassWarning = false)
         {
-            if (!bypassWarning && externalLinkWarning.Value)
+            if (!bypassWarning && externalLinkWarning.Value && dialogOverlay != null)
                 dialogOverlay.Push(new ExternalLinkDialog(url, () => host.OpenUrlExternally(url), () => host.GetClipboard()?.SetText(url)));
             else
                 host.OpenUrlExternally(url);
+        }
+
+        public class ExternalLinkDialog : PopupDialog
+        {
+            public ExternalLinkDialog(string url, Action openExternalLinkAction, Action copyExternalLinkAction)
+            {
+                HeaderText = "Just checking...";
+                BodyText = $"You are about to leave osu! and open the following link in a web browser:\n\n{url}";
+
+                Icon = FontAwesome.Solid.ExclamationTriangle;
+
+                Buttons = new PopupDialogButton[]
+                {
+                    new PopupDialogOkButton
+                    {
+                        Text = @"Yes. Go for it.",
+                        Action = openExternalLinkAction
+                    },
+                    new PopupDialogCancelButton
+                    {
+                        Text = @"Copy URL to the clipboard instead.",
+                        Action = copyExternalLinkAction
+                    },
+                    new PopupDialogCancelButton
+                    {
+                        Text = @"No! Abort mission!"
+                    },
+                };
+            }
         }
     }
 }
