--- osu.Game/Overlays/Chat/ExternalLinkDialog.cs
@@ -1,40 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-#nullable disable
-
-using System;
-using osu.Framework.Graphics.Sprites;
-using osu.Game.Overlays.Dialog;
-
-namespace osu.Game.Overlays.Chat
-{
-    public class ExternalLinkDialog : PopupDialog
-    {
-        public ExternalLinkDialog(string url, Action openExternalLinkAction, Action copyExternalLinkAction)
-        {
-            HeaderText = "Just checking...";
-            BodyText = $"You are about to leave osu! and open the following link in a web browser:\n\n{url}";
-
-            Icon = FontAwesome.Solid.ExclamationTriangle;
-
-            Buttons = new PopupDialogButton[]
-            {
-                new PopupDialogOkButton
-                {
-                    Text = @"Yes. Go for it.",
-                    Action = openExternalLinkAction
-                },
-                new PopupDialogCancelButton
-                {
-                    Text = @"Copy URL to the clipboard instead.",
-                    Action = copyExternalLinkAction
-                },
-                new PopupDialogCancelButton
-                {
-                    Text = @"No! Abort mission!"
-                },
-            };
-        }
-    }
-}
